creating validate current user

+++ src/controllers/ActivityController.java
@@ -2,6 +2,7 @@ package controllers;
 
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collections;
 import java.util.List;
 
 import models.activity.Activity;
@@ -15,48 +16,107 @@ public class ActivityController {
         activity.save();
     }
 
-    public void update(Activity activity) throws Exception {
+    public void update(Activity activity, int userId) throws Exception {
+        if (!isOwner(activity, userId)) {
+            throw new Exception("Permissao negada");
+        }
         activity.update();
     }
 
-    public void delete(Activity activity) throws Exception {
+    public void delete(Activity activity, int userId) throws Exception {
+        if (!isOwner(activity, userId)) {
+            throw new Exception("Permissao negada");
+        }
         activity.delete();
     }
 
-    public Activity findById(int id) throws Exception {
+    public Activity findById(int id, int userId) throws Exception {
         Activity searchedActivity;
 
         searchedActivity = LeisureActivity.findById(id);
-        if (searchedActivity != null) return searchedActivity;
+        if (searchedActivity != null && isOwner(searchedActivity, userId)) return searchedActivity;
 
         searchedActivity = PhysicalActivity.findById(id);
-        if (searchedActivity != null) return searchedActivity;
+        if (searchedActivity != null && isOwner(searchedActivity, userId)) return searchedActivity;
 
         searchedActivity = WorkActivity.findById(id);
-        if (searchedActivity != null) return searchedActivity;
+        if (searchedActivity != null && isOwner(searchedActivity, userId)) return searchedActivity;
         
         throw new Exception("Nenhum registro encontrado");
     }
 
-    public List<Activity> findAll() throws Exception {
+    public List<Activity> findAll(int userId) throws Exception {
         List<Activity> activities = new ArrayList<Activity>();
 
         activities.addAll(PhysicalActivity.findAll());
         activities.addAll(LeisureActivity.findAll());
         activities.addAll(WorkActivity.findAll());
 
+        for (Activity activity: activities) {
+            if (!isOwner(activity, userId)) activities.remove(activity);
+        }
+
         return activities;      
     }
 
 
-    public List<Activity> findByDate(Calendar initialDate, Calendar finalDate) throws Exception {
+    public List<Activity> findByDate(Calendar initialDate, Calendar finalDate, int userId) throws Exception {
         List<Activity> activities = new ArrayList<Activity>();
 
         activities.addAll(LeisureActivity.findByDate(initialDate, finalDate));
         activities.addAll(PhysicalActivity.findByDate(initialDate, finalDate));
         activities.addAll(WorkActivity.findByDate(initialDate, finalDate));
 
-        return activities;
+        for (Activity activity: activities) {
+            if (!isOwner(activity, userId)) activities.remove(activity);
+        }
 
+        return activities;
+    }
+
+
+    public List<Activity> findByCategory(int activityCategory, int userId) throws Exception {
+        List<Activity> activities = new ArrayList<Activity>();
+
+
+        switch (activityCategory) {
+            case 1: // FISICA
+                activities.addAll(PhysicalActivity.findAll());
+                break;
+            case 2: //LAZER
+                activities.addAll(LeisureActivity.findAll());
+                break;
+            case 3: // TRABALHO
+                activities.addAll(WorkActivity.findAll());
+                break;
+            default:
+                throw new Exception("Informe uma opcao valida");
+        }
+
+        for (Activity activity: activities) {
+            if (!isOwner(activity, userId)) activities.remove(activity);
+        }
+
+        return activities;
+    }
+
+
+    public List<Activity> ranking(int userId) throws Exception {
+        List<Activity> activities = this.findAll(userId);       
+         
+        for (Activity activity: activities) {
+            if (!isOwner(activity, userId)) activities.remove(activity);
+        }
+
+        Collections.sort(activities);
+        Collections.reverse(activities);
+
+
+        return activities;
+    }
+
+    private static boolean isOwner(Activity activity, int userId) throws Exception {
+        if (activity.getOwner() == userId) return true;
+        return false;
     }
 }
